Add custom message lambda to user pools
