+++ backend/compact-connect/lambdas/nodejs/cognito-emails/handler.ts
@@ -3,11 +3,10 @@ import { SESClient } from '@aws-sdk/client-ses';
 import { S3Client } from '@aws-sdk/client-s3';
 import { Lambda } from './lambda';
 
-
 const lambda = new Lambda({
     dynamoDBClient: new DynamoDBClient(),
     s3Client: new S3Client(),
     sesClient: new SESClient(),
 });
 
-export const sendEmail = lambda.handler.bind(lambda);
+export const customMessage = lambda.handler.bind(lambda);
