+++ backend/compact-connect/lambdas/nodejs/jest.config.mjs
@@ -2,7 +2,7 @@ export default {
     preset: 'ts-jest',
     transform: {}, // Disables all transformations to commonJS
     testEnvironment: 'node',
-    testMatch: ['**/tests/*.test.[jt]s?(x)'],
+    testMatch: ['**/tests/**/*.test.[jt]s?(x)'],
     moduleFileExtensions: ['ts', 'js'],
     verbose: true,
     testPathIgnorePatterns: [
