+++ backend/compact-connect/lambdas/nodejs/lib/email/index.ts
@@ -1,0 +1,3 @@
+export { CognitoEmailService } from './cognito-email-service';
+export { EmailNotificationService } from './email-notification-service';
+export { IngestEventEmailService } from './ingest-event-email-service';
