+++ backend/compact-connect/lambdas/nodejs/tests/lib/email/cognito-email-service.test.ts
@@ -1,0 +1,114 @@
+import { mockClient } from 'aws-sdk-client-mock';
+import 'aws-sdk-client-mock-jest';
+import { Logger } from '@aws-lambda-powertools/logger';
+import { SESClient } from '@aws-sdk/client-ses';
+import { CognitoEmailService } from '../../../lib/email';
+import { describe, it, expect, beforeEach, jest } from '@jest/globals';
+
+const asSESClient = (mock: ReturnType<typeof mockClient>) =>
+    mock as unknown as SESClient;
+
+describe('CognitoEmailService', () => {
+    let emailService: CognitoEmailService;
+    let mockSESClient: ReturnType<typeof mockClient>;
+
+    beforeEach(() => {
+        jest.clearAllMocks();
+        mockSESClient = mockClient(SESClient);
+
+        // Reset environment variables
+        process.env.FROM_ADDRESS = 'noreply@example.org';
+        process.env.UI_BASE_PATH_URL = 'https://app.test.compactconnect.org';
+
+        emailService = new CognitoEmailService({
+            logger: new Logger({ serviceName: 'test' }),
+            sesClient: asSESClient(mockSESClient),
+            s3Client: {} as any,
+            compactConfigurationClient: {} as any,
+            jurisdictionClient: {} as any
+        });
+    });
+
+    describe('generateCognitoEmailTemplate', () => {
+        it('should generate HTML email template with header and footer', () => {
+            const template = emailService.generateCognitoEmailTemplate(
+                'Test Subject',
+                'Test body text'
+            );
+
+            expect(template).toContain('<!DOCTYPE html');
+            expect(template).toContain('Test body text');
+            expect(template).toContain('CompactConnect');
+        });
+    });
+
+    describe('generateCognitoMessage', () => {
+        it('should generate AdminCreateUser message', () => {
+            const { subject, htmlContent } = emailService.generateCognitoMessage(
+                'CustomMessage_AdminCreateUser',
+                '{####}',
+                'testuser'
+            );
+
+            expect(subject).toBe('Welcome to CompactConnect');
+            expect(htmlContent).toContain('Your temporary password is: {####}');
+            expect(htmlContent).toContain('Username: testuser');
+            expect(htmlContent).toContain('https://app.test.compactconnect.org/Login');
+        });
+
+        it('should generate ForgotPassword message', () => {
+            const { subject, htmlContent } = emailService.generateCognitoMessage(
+                'CustomMessage_ForgotPassword',
+                '{####}'
+            );
+
+            expect(subject).toBe('Reset your password');
+            expect(htmlContent).toContain('Enter the following code to proceed: {####}');
+        });
+
+        it('should generate UpdateUserAttribute message', () => {
+            const { subject, htmlContent } = emailService.generateCognitoMessage(
+                'CustomMessage_UpdateUserAttribute',
+                '{####}'
+            );
+
+            expect(subject).toBe('Verify your email');
+            expect(htmlContent).toContain('Please verify your new email address by entering the following code: {####}');
+        });
+
+        it('should generate VerifyUserAttribute message', () => {
+            const { subject, htmlContent } = emailService.generateCognitoMessage(
+                'CustomMessage_VerifyUserAttribute',
+                '{####}'
+            );
+
+            expect(subject).toBe('Verify your email');
+            expect(htmlContent).toContain('Please verify your email address by entering the following code: {####}');
+        });
+
+        it('should generate ResendCode message', () => {
+            const { subject, htmlContent } = emailService.generateCognitoMessage(
+                'CustomMessage_ResendCode',
+                '{####}'
+            );
+
+            expect(subject).toBe('New verification code for CompactConnect');
+            expect(htmlContent).toContain('Your new verification code is: {####}');
+        });
+
+        it('should generate SignUp message', () => {
+            const { subject, htmlContent } = emailService.generateCognitoMessage(
+                'CustomMessage_SignUp',
+                '{####}'
+            );
+
+            expect(subject).toBe('Welcome to CompactConnect');
+            expect(htmlContent).toContain('Please verify your email address by entering the following code: {####}');
+        });
+
+        it('should throw error for unsupported trigger source', () => {
+            expect(() => emailService.generateCognitoMessage('UnsupportedTrigger'))
+                .toThrow('Unsupported Cognito trigger source: UnsupportedTrigger');
+        });
+    });
+});
