+++ backend/compact-connect/lambdas/nodejs/tests/lib/event-client.test.ts
@@ -1,7 +1,7 @@
 import { Logger } from '@aws-lambda-powertools/logger';
 import { DynamoDBClient, QueryCommand } from '@aws-sdk/client-dynamodb';
 import { mockClient } from 'aws-sdk-client-mock';
-import { EventClient } from '../lib/event-client';
+import { EventClient } from '../../lib/event-client';
 import { describe, it, expect, beforeAll, beforeEach, jest } from '@jest/globals';
